move interface - CsvLineConvertible
